Add 128 bit integers, mapped to Arrow decimal. Parser doesn't allow the most negative value as it parses first and negates after. Still overflows to double for now.

+++ lib/endb_lib/endb.h
@@ -150,7 +150,7 @@ typedef struct Ast {
       enum Keyword kw;
     };
     struct {
-      int64_t integer;
+      i128 integer;
     };
     struct {
       double float_;
